My part of the final commit for IG-4621, which includes the new OAuth login file, the modified .gitignore, and the changes to the skyapi-testing files. The skyapi-testing files were renamed to be in a "Retired" folder, and the index.js file was modified to reflect the new file paths. The OAuth_Login copy file is not staged for commit because it has both modified content and untracked content, so it will need to be reviewed and staged separately before it can be committed.
