+++ .gitignore
@@ -7,3 +7,5 @@
 debug.log
 node_modules
 .env
+/package-lock.json
+/package.json
